Move common continual block into models.common
