+++ models/common/__init__.py
@@ -1,0 +1,1 @@
+from .res import CoResBlock, CoResStage, c2_msra_fill, init_weights  # noqa: F401
